RF transformations
Included all main matrix transformations relevant for deembedding as
methods of the measurement class.
Creating the y matrices as an attribute requires now execution of the
create_y method.

Fixes #6 .
For the future, the transformation "matrix of vectors to vector of
matrices" can be done via
np.array([mat[:,:,i] for i in range(len(self.freq))])