Changed init function and changed deembedding of through
Now, a measurement instance can be created either by passing a file (in
this case nothing changes compared to the older version) or by passing
the "from_s" together with the "parent_meas" arguments. In this case,
simply pass calculated s parameters as "from_s" and another measurement
instance as "parent_meas" as arguments.